[Refactor] Rename package from qube to pyqube
