+++ examples/rest_example.py
@@ -1,5 +1,5 @@
 import os
-from qube import QubeClient
+from pyqube import QubeClient
 
 # Retrieve required environment variables for API configuration
 API_KEY = os.environ.get(
