+++ pyqube/events/clients.py
@@ -2,8 +2,8 @@ import paho.mqtt.client as mqtt
 from datetime import UTC, datetime
 from typing import Callable, Dict, Optional
 
-from qube.events.exceptions import MessageHandlingError, SubscriptionError
-from qube.events.handlers import QueueHandler, QueuingSystemResetHandler, TicketHandler
+from pyqube.events.exceptions import MessageHandlingError, SubscriptionError
+from pyqube.events.handlers import QueueHandler, QueuingSystemResetHandler, TicketHandler
 
 
 class MQTTClient(TicketHandler, QueuingSystemResetHandler, QueueHandler):
