+++ pyqube/events/exceptions.py
