+++ pyqube/events/tests/test_queue_handler.py
@@ -1,8 +1,8 @@
 import unittest
 from unittest.mock import MagicMock, ANY, Mock
 
-from qube.events.handlers import QueueHandler
-from qube.types import QueueWithAverageWaitingTime, QueueWithWaitingTickets, QueueGeneralDetails
+from pyqube.events.handlers import QueueHandler
+from pyqube.types import QueueWithAverageWaitingTime, QueueWithWaitingTickets, QueueGeneralDetails
 
 
 class ConcreteQueueHandler(QueueHandler):
