+++ pyqube/events/tests/test_queuing_system_reset_decorator.py
@@ -1,7 +1,7 @@
 import pytest
 import json
 from unittest.mock import Mock
-from qube.types import QueuingSystemReset
+from pyqube.types import QueuingSystemReset
 
 
 class TestQueuingSystemResetDecorator:
@@ -14,7 +14,7 @@ class TestQueuingSystemResetDecorator:
         """
         Fixture to create an instance of MQTTClient.
         """
-        from qube.events.clients import MQTTClient
+        from pyqube.events.clients import MQTTClient
         return MQTTClient(api_key="dummy_api_key", location_id=1)
 
     def simulate_mqtt_message(self, mqtt_client, topic, payload):
