+++ pyqube/rest/tests/test_list_queues_of_queues_list.py
@@ -3,10 +3,10 @@ import unittest
 from unittest import mock
 from unittest.mock import call, patch
 
-from qube.rest.clients import RestClient
-from qube.rest.exceptions import BadRequest, Forbidden, NotAuthorized, NotFound
-from qube.rest.graphql_generators import QueuesListGraphQLGenerator
-from qube.types import Queue
+from pyqube.rest.clients import RestClient
+from pyqube.rest.exceptions import BadRequest, Forbidden, NotAuthorized, NotFound
+from pyqube.rest.graphql_generators import QueuesListGraphQLGenerator
+from pyqube.types import Queue
 
 
 @patch.object(RestClient, "make_graphql_request")
