+++ pyqube/rest/tests/test_set_current_counter.py
@@ -2,15 +2,15 @@ import unittest
 from unittest import mock
 from unittest.mock import patch
 
-from qube.rest.clients import RestClient
-from qube.rest.exceptions import (
+from pyqube.rest.clients import RestClient
+from pyqube.rest.exceptions import (
     BadRequest,
     Forbidden,
     NoAccessToCounterException,
     NotAuthorized,
     NotFound,
 )
-from qube.types import LocationAccessWithCurrentCounter
+from pyqube.types import LocationAccessWithCurrentCounter
 
 
 @patch.object(RestClient, "put_request")
