+++ pyqube/rest/tests/test_set_queue_status.py
@@ -2,15 +2,15 @@ import unittest
 from unittest import mock
 from unittest.mock import patch
 
-from qube.rest.clients import RestClient
-from qube.rest.exceptions import (
+from pyqube.rest.clients import RestClient
+from pyqube.rest.exceptions import (
     BadRequest,
     Forbidden,
     NoAccessToCounterException,
     NotAuthorized,
     NotFound,
 )
-from qube.types import Queue
+from pyqube.types import Queue
 
 
 @patch.object(RestClient, "put_request")
